refactor: 调整组件目录，方便后续功能开发
