+++ src/components/Base/Modal/index.js
@@ -1,0 +1,3 @@
+import Modal from './index.vue';
+
+export default Modal;
