+++ src/components/InteractionsReceive/Announcement/Entrance.vue
