+++ src/components/InteractionsReceive/Announcement/MobileAnnouncement.vue
