+++ src/components/InteractionsReceive/Announcement/PCAnnouncement.vue
