+++ src/components/InteractionsReceive/Announcement/index.js
