+++ src/components/InteractionsReceive/AnswerCard/MobileAnswerCard.vue
