+++ src/components/InteractionsReceive/Lottery/Entrance.vue
