+++ src/components/InteractionsReceive/Lottery/MobileLottery.vue
