+++ src/components/InteractionsReceive/Lottery/PCLottery.vue
