+++ src/components/InteractionsReceive/Lottery/index.js
