+++ src/components/InteractionsReceive/PushCard/MobilePushCard.vue
