+++ src/components/InteractionsReceive/PushCard/PCPushCard.vue
