+++ src/components/InteractionsReceive/Questionnaire/Entrance.vue
