+++ src/components/InteractionsReceive/Questionnaire/MobileQuestionnaire.vue
