+++ src/components/InteractionsReceive/WelfareLottery/Entrance.vue
