+++ src/components/InteractionsReceive/WelfareLottery/MobileWelfareLottery.vue
