+++ src/components/InteractionsReceive/WelfareLottery/PCWelfareLottery.vue
