+++ src/components/InteractionsReceive/WelfareLottery/index.js
