+++ src/plugins/vant.js
@@ -1,6 +1,6 @@
+// 借助 babel-plugin-import 自动导入到 vue 中
 import { Dialog, Toast } from 'vant';
 
-// 借助 babel-plugin-import 自动导入到 vue 中
 Toast.setDefaultOptions({
   duration: 3 * 1000
 });
